src/com/duncangb/islands/terrain/Coordinate.java
package com.duncangb.islands.terrain;

public class Coordinate {
    private final int x;
    private final int y;

    public Coordinate(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public int getX() {
        return this.x;
    }

    public int getY() {
        return this.y;
    }

    @Override
    public int hashCode() {
        return 31 * x + y;
    }

    @Override
    public boolean equals(Object o) {
        if (o instanceof Coordinate) {
            Coordinate other = (Coordinate) o;
            return other.getX() == this.getX() && other.getY() == this.getY();
        } else {
            return false;
        }
    }
}
